Lift Loader.parse into System.parse. This allows Loader to be independent of compiler/transcoder and it allows the System to control parse options w/o consulting Loader.
